chore: Refactor CPU execution of operators
